requirements.txt
alembic==1.8.0
APIFlask==1.1.2
apispec==5.2.2
astroid==2.11.6
attrs==21.4.0
bcrypt==3.2.2
black==22.3.0
blinker==1.5
CacheControl==0.12.11
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
coverage==6.4.2
cyclonedx-python-lib==2.7.1
dill==0.3.5.1
Faker==13.12.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
mccabe==0.7.0
msgpack==1.0.4
mypy-extensions==0.4.3
packageurl-python==0.10.1
packaging==21.3
pathspec==0.9.0
pip-api==0.0.30
pip-requirements-parser==31.2.0
pip_audit==2.4.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pylint==2.14.4
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
resolvelib==0.8.1
rich==12.5.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.37
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
urllib3==1.26.11
webargs==8.1.0
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
